requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
environs==14.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
inflection==0.5.1
marshmallow==4.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
